requirements.txt
requests
flask
Flask-PyMongo
flask-cors
mongomock
gunicorn
jsonpatch
PyJWT
